requirements.txt
streamlit
pandas
scikit-learn
nltk
python-docx
PyMuPDF
